input-device: move input events definition into objects
Keep the consistent hierachy
