+++ include/taiwins/input_device.h
@@ -25,6 +25,8 @@
 #include <stdint.h>
 #include <wayland-server.h>
 #include <xkbcommon/xkbcommon.h>
+#include <taiwins/objects/seat_grab.h>
+#include <taiwins/objects/gestures.h>
 
 #ifdef  __cplusplus
 extern "C" {
@@ -158,102 +160,6 @@ struct tw_input_device {
 };
 
 struct tw_output_device;
-
-/** keyboard event */
-struct tw_event_keyboard_key {
-	struct tw_input_device *dev;
-	uint32_t time;
-	uint32_t keycode;
-	enum wl_keyboard_key_state state;
-};
-
-struct tw_event_keyboard_modifier {
-	struct tw_input_device *dev;
-	uint32_t depressed, latched, locked, group;
-};
-
-struct tw_event_keyboard_keymap {
-	struct tw_input_device *dev;
-	const struct xkb_keymap *keymap;
-};
-
-/** pointer events */
-struct tw_event_pointer_button {
-	struct tw_input_device *dev;
-	enum wl_pointer_button_state state;
-	uint32_t button;
-	uint32_t time;
-};
-
-struct tw_event_pointer_motion {
-	struct tw_input_device *dev;
-	uint32_t time;
-	double delta_x, delta_y;
-	double unaccel_dx, unaccel_dy;
-};
-
-struct tw_event_pointer_motion_abs {
-	struct tw_input_device *dev;
-	const struct tw_output_device *output;
-	uint32_t time_msec;
-	// From 0..1
-	double x, y;
-};
-
-struct tw_event_pointer_axis {
-	struct tw_input_device *dev;
-	uint32_t time;
-	enum wl_pointer_axis_source source;
-	enum wl_pointer_axis axis;
-	double delta;
-	int32_t delta_discrete;
-};
-
-struct tw_event_pointer_gesture {
-	struct tw_input_device *dev;
-	uint32_t time, fingers;
-
-	enum {
-		TW_POINTER_GESTURE_BEGIN = 0,
-		TW_POINTER_GESTURE_UPDATE,
-		TW_POINTER_GESTURE_END,
-	} state;
-
-	double dx, dy, rotation, scale;
-	bool cancelled;
-};
-
-/** touch events */
-struct tw_event_touch_down {
-	struct tw_input_device *dev;
-	const struct tw_output_device *output;
-	uint32_t time;
-	int32_t touch_id;
-	// From 0..1
-	double x, y;
-};
-
-struct tw_event_touch_up {
-	struct tw_input_device *dev;
-	uint32_t time;
-	int32_t touch_id;
-};
-
-struct tw_event_touch_motion {
-	struct tw_input_device *dev;
-	const struct tw_output_device *output;
-
-	uint32_t time;
-	int32_t touch_id;
-	// From 0..1
-	double x, y;
-};
-
-struct tw_event_touch_cancel {
-	struct tw_input_device *dev;
-	uint32_t time_msec;
-	int32_t touch_id;
-};
 
 /**
  * Emit input signals
