+++ include/taiwins/objects/seat_grab.h
@@ -32,6 +32,23 @@ extern "C" {
  * keyboard
  *****************************************************************************/
 
+struct tw_event_keyboard_key {
+	struct tw_input_device *dev;
+	uint32_t time;
+	uint32_t keycode;
+	enum wl_keyboard_key_state state;
+};
+
+struct tw_event_keyboard_modifier {
+	struct tw_input_device *dev;
+	uint32_t depressed, latched, locked, group;
+};
+
+struct tw_event_keyboard_keymap {
+	struct tw_input_device *dev;
+	int fd; uint32_t format, size;
+};
+
 void
 tw_keyboard_default_enter(struct tw_seat_keyboard_grab *grab,
                           struct wl_resource *surface, uint32_t *keycodes,
@@ -80,6 +97,38 @@ tw_keyboard_notify_modifiers(struct tw_keyboard *keyboard,
 /******************************************************************************
  * pointer
  *****************************************************************************/
+
+/** pointer events */
+struct tw_event_pointer_button {
+	struct tw_input_device *dev;
+	enum wl_pointer_button_state state;
+	uint32_t button;
+	uint32_t time;
+};
+
+struct tw_event_pointer_motion {
+	struct tw_input_device *dev;
+	uint32_t time;
+	double delta_x, delta_y;
+	double unaccel_dx, unaccel_dy;
+};
+
+struct tw_event_pointer_motion_abs {
+	struct tw_input_device *dev;
+	const struct tw_output_device *output;
+	uint32_t time_msec;
+	// From 0..1
+	double x, y;
+};
+
+struct tw_event_pointer_axis {
+	struct tw_input_device *dev;
+	uint32_t time;
+	enum wl_pointer_axis_source source;
+	enum wl_pointer_axis axis;
+	double delta;
+	int32_t delta_discrete;
+};
 
 void
 tw_pointer_default_enter(struct tw_seat_pointer_grab *grab,
@@ -148,6 +197,37 @@ tw_pointer_notify_frame(struct tw_pointer *pointer)
 /******************************************************************************
  * touch
  *****************************************************************************/
+
+struct tw_event_touch_down {
+	struct tw_input_device *dev;
+	const struct tw_output_device *output;
+	uint32_t time;
+	int32_t touch_id;
+	// From 0..1
+	double x, y;
+};
+
+struct tw_event_touch_up {
+	struct tw_input_device *dev;
+	uint32_t time;
+	int32_t touch_id;
+};
+
+struct tw_event_touch_motion {
+	struct tw_input_device *dev;
+	const struct tw_output_device *output;
+
+	uint32_t time;
+	int32_t touch_id;
+	// From 0..1
+	double x, y;
+};
+
+struct tw_event_touch_cancel {
+	struct tw_input_device *dev;
+	uint32_t time_msec;
+	int32_t touch_id;
+};
 
 void
 tw_touch_default_down(struct tw_seat_touch_grab *grab, uint32_t time_msec,
